+++ api/transport/transporttest/reqres.go
@@ -83,6 +83,13 @@ func (m RequestMatcher) Matches(got interface{}) bool {
 		return false
 	}
 
+	// We check if 'Transport' is set before comparing, since transports may
+	// modify the field after users define the struct.
+	if l.Transport != "" && l.Transport != r.Transport {
+		m.t.Logf("Transport mismatch: %s != %s", l.Transport, r.Transport)
+		return false
+	}
+
 	if l.Encoding != r.Encoding {
 		m.t.Logf("Encoding mismatch: %s != %s", l.Service, r.Service)
 		return false